[util] Move text options into text_options_t
Remove now empty options.cc.
