+++ util/hb-ot-shape-closure.cc
@@ -26,6 +26,7 @@
 
 #include "shape-options.hh"
 #include "font-options.hh"
+#include "text-options.hh"
 #include "main-font-text.hh"
 
 const unsigned DEFAULT_FONT_SIZE = FONT_SIZE_NONE;
