+++ util/hb-subset.cc
@@ -30,6 +30,7 @@
 #include "subset-options.hh"
 #include "output-options.hh"
 #include "face-options.hh"
+#include "text-options.hh"
 #include "main-font-text.hh"
 
 /*
